refactor: streamline fetching and caching of followed streams from Twitch and YouTube; consolidate logic into helper functions
